:recycle: Refactor: Service 클래스 세분화 및 컨트롤러 수정
기존에는 StudyGroupService에 모든 기능의 로직까지 포함하는 형태였는데,
이를 별도로 기능적으로 구분해서 Service와 Controller를 세분화하도록 함

- 스터디그룹은 생성 / 팀원 관리 사이드
- 데일리 로그 / 공지사항 / 일정은 별도로 구분한다.
