+++ BE_SIDE/study/src/main/java/sw/study/studyGroup/service/StudyGroupService.java
@@ -452,67 +452,4 @@ public class StudyGroupService {
                 }
         );
     }
-
-    // 데일리 로그 작성
-    public void createDailyLog(String accessToken, long groupId, String title, String content){
-
-        Member member = currentLogginedInfo(accessToken);
-
-        Participant participant = participantRepository.findByMemberIdAndStudyGroupId(member.getId(), groupId)
-                .orElseThrow(() -> new UnauthorizedException("해당 그룹에 참가하지 않은 비정상적인 접근입니다."));
-
-        StudyGroup studyGroup = studyGroupRepository.findById(groupId)
-                .orElseThrow(() -> new StudyGroupNotFoundException("스터디 그룹이 존재하지 않습니다."));
-
-        DailyLog dailyLog = DailyLog.createDailyLog(studyGroup, participant, title, content);
-        dailyLogRepository.save(dailyLog);
-    }
-
-    // 데일리 로그 조회
-    @Transactional(readOnly = true)
-    public List<DailyLogResponseDto> listOfDailyLog(String accessToken, long groupId){
-        
-        Member member = currentLogginedInfo(accessToken);
-
-        participantRepository.findByMemberIdAndStudyGroupId(member.getId(), groupId)
-                .orElseThrow(() -> new UnauthorizedException("해당 그룹에 참가하지 않은 비정상적인 접근입니다."));
-
-        List<DailyLog> logs = dailyLogRepository.findAllByStudyGroupId(groupId);
-        return logs.stream().map(DailyLogResponseDto::new).toList();
-        // 스트림 형태로 변환, 각 요소를 Dto 형태로 바꿔서 리턴
-    }
-
-    // 데일리 로그 수정
-    public void updateDailyLog(String accessToken, long groupId, long logId, String title,String content){
-        Member member = currentLogginedInfo(accessToken);
-
-        Participant participant = participantRepository.findByMemberIdAndStudyGroupId(member.getId(), groupId)
-                .orElseThrow(() -> new UnauthorizedException("해당 그룹에 참가하지 않은 비정상적인 접근입니다."));
-
-        DailyLog dailyLog =dailyLogRepository.findById(logId)
-                .orElseThrow(()-> new DailyLogNotFoundException("해당하는 데일리 로그는 존재하지 않습니다."));
-
-        if(!dailyLog.getAuthor().getId().equals(participant.getId()))
-            throw new UnauthorizedException("수정 권한이 없습니다.");
-
-        dailyLog.updateLog(title, content);
-    }
-
-    // 데일리 로그 삭제
-    public void deleteDailyLog(String accessToken, long groupId, long logId){
-        Member member = currentLogginedInfo(accessToken);
-
-        Participant participant = participantRepository.findByMemberIdAndStudyGroupId(member.getId(), groupId)
-                .orElseThrow(() -> new UnauthorizedException("해당 그룹에 참가하지 않은 비정상적인 접근입니다."));
-
-        DailyLog dailyLog =dailyLogRepository.findById(logId)
-                .orElseThrow(()-> new DailyLogNotFoundException("해당하는 데일리 로그는 존재하지 않습니다."));
-
-        if (dailyLog.getAuthor().getId().equals(participant.getId()) || participant.getRole() == Role.LEADER) {
-            // 작성자 본인이거나 혹은 방장일 경우에
-            dailyLogRepository.deleteById(logId);
-        } else {
-            throw new UnauthorizedException("삭제 권한이 없습니다.");
-        }
-    }
 }
